% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assorthead-package.R
\docType{package}
\name{assorthead-package}
\alias{assorthead}
\alias{assorthead-package}
\title{assorthead: Assorted Header-Only C++ Libraries}
\description{
Vendors an assortment of useful header-only C++ libraries. Bioconductor packages can use these libraries in their own C++ code by LinkingTo this package without introducing any additional dependencies. The use of a central repository avoids duplicate vendoring of libraries across multiple R packages, and enables better coordination of version updates across cohorts of interdependent C++ libraries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LTLA/assorthead}
  \item Report bugs at \url{https://github.com/LTLA/assorthead/issues}
}

}
\author{
\strong{Maintainer}: Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com}

}
